使用common.ps1來複用部分代碼。

+++ pack.ps1
@@ -14,6 +14,7 @@ $file = Get-Path
 if ($file -eq "") {
     $file = "file"
 }
+
 if ($args.Count -ge 1) {
     $num = $args[0]
 }
@@ -22,32 +23,36 @@ if ($args.Count -ge 2) {
 }
 Remove-Item *.patch
 git format-patch $num
-if (Test-Path logseq.rar) {
-    Remove-Item logseq.rar
+if (Test-Path ($file + ".rar")) {
+    Remove-Item ($file + ".rar")
 }
 
-$h = ""
+# 导入rar命令
+common.ps1
 
-if (Test-Path "$PSScriptRoot/password.txt") {
-    $pw = Get-Content "$PSScriptRoot/password.txt"
-    $h = "-hp" + $pw
-}
+# 獲取密碼
+$h = get_password
 
 # 阻塞運行
-$process = Start-Process -FilePath WinRAR -ArgumentList "a ${file} *.patch -m5 ${h} -T" -NoNewWindow -Wait -PassThru
-Write-Output $process.ExitCode # 進程返回值
-if ($process.ExitCode -ne 0) {
+if (winrar_compress $file $h *.patch -ne 0) {
     Write-Output "壓縮失敗"
     exit 1
 }
 Remove-Item *.patch
+if ($LASTEXITCODE -ne 0) {
+    Write-Output "刪除失敗"
+    exit 1
+}
 
 # 打包後複製到指定位置
-$config = "$PSScriptRoot/config.json"
-if (Test-Path $config) {
-    $json = Get-Content $config -Raw | ConvertFrom-Json
-    if ($json.path -and (Test-Path $json.path)) {
+$json = get_config
+if ($json -and $json.path -and (Test-Path $json.path)) {
+    Copy-Item -Path *.rar -Destination $json.path
+    if ($LASTEXITCODE -eq 0) {
         Write-Output ("複製到" + $json.path)
-        Copy-Item -Path *.rar -Destination $json.path
+        Remove-Item -Path *.rar
+        if ($LASTEXITCODE -ne 0) {
+            Write-Output "刪除失敗"
+        }
     }
 }
